tests: Move token tests to their own module
